+++ src/main/java/gui/AddDistantLightFrame.java
@@ -63,7 +63,7 @@ class AddDistantLightFrame extends AddLightSourceFrame {
     }
 
     @Override
-    LightSource getParticularLightSource(double intensity, Vector3f color, Vector3f xyz) {
+    LightSource getParticularLightSource(double intensity, Vector3d color, Vector3d xyz) {
         //remapping intensity over [0,10]
         intensity *= 10;
         if (removeMode) {
