+++ src/main/java/gui/AddPointLightFrame.java
@@ -65,7 +65,7 @@ class AddPointLightFrame extends AddLightSourceFrame {
 
 
     @Override
-    LightSource getParticularLightSource(double intensity, Vector3f color, Vector3f xyz) {
+    LightSource getParticularLightSource(double intensity, Vector3d color, Vector3d xyz) {
         //remapping intensity over [0,150]
         intensity *= 150;
         if (removeMode) {
