+++ src/main/java/gui/ControlPointsSurfaceFrame.java
@@ -2,9 +2,6 @@ package gui;
 
 import rendering.*;
 
-import javax.swing.*;
-import java.awt.*;
-
 abstract class ControlPointsSurfaceFrame extends AddObjectFrame {
 
 
@@ -17,10 +14,10 @@ abstract class ControlPointsSurfaceFrame extends AddObjectFrame {
         super(visualizer, scene, defaultSceneObject);
     }
 
-    abstract void updateControlPoints(Vector3f[][] cp);
+    abstract void updateControlPoints(Vector3d[][] cp);
 
 
-    public void showControlPoints(Vector3f[][] cp, Matrix4D otw) {
+    public void showControlPoints(Vector3d[][] cp, Matrix4D otw) {
         for (int i = 0; i <cp.length; i++) {
             for (int j = 0; j < cp[0].length; j++) {
                 PhysicalPoint physicalPoint = new PhysicalPoint(otw.transformVector(cp[i][j]));
