+++ src/main/java/gui/Main.java
@@ -1,7 +1,7 @@
 package gui;
 
 import rendering.Scene;
-import rendering.Vector3f;
+import rendering.Vector3d;
 
 import java.awt.image.BufferedImage;
 
@@ -13,10 +13,10 @@ public class Main {
         int width = 1024;
         int height = 650;
         double fieldOfView = 39.6; //in degrees
-        Vector3f cameraPosition = new Vector3f(0,0,0);
+        Vector3d cameraPosition = new Vector3d(0,0,0);
         BufferedImage img = new BufferedImage(width, height, TYPE_INT_RGB);
 
-        Scene scene = new Scene(width, height, fieldOfView, img, cameraPosition, new Vector3f(0.1f, 0.3f, 0.6f));
+        Scene scene = new Scene(width, height, fieldOfView, img, cameraPosition, new Vector3d(0.1f, 0.3f, 0.6f));
 
         scene.setBVH();
         scene.render(20);
