+++ src/main/java/rendering/BezierCurve2.java
@@ -7,25 +7,25 @@ public class BezierCurve2 {
     //is useful for calculating derivatives
     //of degree 3 curves
 
-    private Vector3f[] controlPoints;
+    private Vector3d[] controlPoints;
 
-    public BezierCurve2(Vector3f[] controlPoints) {
+    public BezierCurve2(Vector3d[] controlPoints) {
         //we should check it contains 4 elements
         this.controlPoints = controlPoints.clone();
     }
 
-    public BezierCurve2(Vector3f p0, Vector3f p1, Vector3f p2) {
-        controlPoints = new Vector3f[4];
+    public BezierCurve2(Vector3d p0, Vector3d p1, Vector3d p2) {
+        controlPoints = new Vector3d[4];
         controlPoints[0] = p0;
         controlPoints[1] = p1;
         controlPoints[2] = p2;
     }
 
-    public Vector3f evaluate(double t) {
-        Vector3f p10 = controlPoints[0].mul(1-t).add(controlPoints[1].mul(t));
-        Vector3f p11 = controlPoints[1].mul(1-t).add(controlPoints[2].mul(t));
+    public Vector3d evaluate(double t) {
+        Vector3d p10 = controlPoints[0].mul(1-t).add(controlPoints[1].mul(t));
+        Vector3d p11 = controlPoints[1].mul(1-t).add(controlPoints[2].mul(t));
 
-        Vector3f p20 = p10.mul(1-t).add(p11.mul(t));
+        Vector3d p20 = p10.mul(1-t).add(p11.mul(t));
 
         return p20;
